+++ examples/api/__tests__/using-jest.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import {
   Conversation,
   WebMessengerGuestSession,
@@ -7,6 +8,7 @@ import {
 // Set realistic timeout value for a Web Messenger conversation
 jest.setTimeout(20000); // 20s
 
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 require('dotenv').config({ path: '../../.env' });
 
 describe('Using jest to perform the test', () => {
