+++ examples/api/src/js-script.js
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+
 require('dotenv').config({ path: '../../.env' });
 
 // << test-section
