+++ examples/api/src/ts-script.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import {
   Conversation,
   WebMessengerGuestSession,
