+++ packages/genesys-web-messaging-tester-cli/src/scenarioCommand/testScript/validateSessionConfig.ts
